requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
pywizlight>=0.5.14
tapo>=0.8.0
pydreo-cloud>=1.0.0
python-dotenv>=1.0.0
eventlet>=0.30.0
